### 2026-07-15 #10962 @editorialbot
[REVIEW]: Bristlecone: an F# library for the long-term ecological detective
**Submitting author:** @AndrewIOM (<a href="http://orcid.org/0000-0001-5540-6161">Andrew C Martin</a>)
**Repository:** https://github.com/AndrewIOM/bristlecone/
**Branch with paper.md** (empty if default branch): master
**Version:** v3.1.2
