### 2026-07-16 AndrewIOM/bristlecone#62 @kMutagene
opening Bristlecone as indicated in the docs not possible
Hi @AndrewIOM, i am currently playing around with this library in the context of the JOSS review by porting some of the [FSharp.Stats growth curve modelling examples](https://fslab.org/FSharp.Stats/GrowthCurve.html), and i found that i cannot open the root namespace in polyglot notebooks:

This is not a big issue ~~as you only show usage in F# scripts in the docs (and in an F# script this works fine), just a heads-up.~~

https://github.com/openjournals/joss-reviews/issues/10962
