### 2026-07-29 AndrewIOM/bristlecone#64 @kMutagene
Paper Review
These items relate to this version of the draft: https://raw.githubusercontent.com/openjournals/joss-papers/joss.10962/joss.10962/10.21105.joss.10962.pdf

Hi @AndrewIOM!

The submission has a clear scope and formulates the state of the field and software design well. Bristlecone is a toolkit for long term ecology modelling, specifically designed as a high-level grammatical layer to formulate hypotheses and handle domain-specific challenges related to the scales of data used for modelling using F# units of measurement. It provides this high level layer for a well-established approach ("the ecological detective") that previously seems to only have been implemented with general-purpose frameworks.

The following are only minor suggestions for revision of the manuscript, mainly concerned with incorrect citation formatting where only the year number, not the author name is included as citation link:

- [ ] L14: Citation formatting seems off
- [x] L49: Missing citations for STAN & JAGS (only stan manual is referenced for a later claim in the text)
- [ ] L58-59: Citation formatting seems off
- [x] L64-66: the F# language could use a citation here, especially since you later mention specific features later on
- [x] L69: Since you maintain the project, I would suggest setting up a zenodo release pipeline for RProvider so you have a stable DOI to reference instead of a website. See for example https://github.com/fable-compiler/Fable/issues/4306
- [x] L122: Citation formatting seems off
- [x] L133: Not immediately clear to me what "custom C" refers to
- [ ] L135: Citation formatting seems off

None of these are showstoppers and should be quick fixes. I will recommend accepting the submission to the editor.

https://github.com/openjournals/joss-reviews/issues/10962
