+++ paper/paper.bib
@@ -154,6 +154,28 @@
   url = {https://mc-stan.org/docs/reference-manual/}
 }
 
+@article{carpenter2017stan,
+  title={{Stan}: A probabilistic programming language},
+  author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+  journal={Journal of Statistical Software},
+  volume={76},
+  number={1},
+  pages={1--32},
+  year={2017},
+  doi={10.18637/jss.v076.i01}
+}
+
+@inproceedings{plummer2003jags,
+  title={{JAGS}: A program for analysis of {B}ayesian graphical models using {G}ibbs sampling},
+  author={Plummer, Martyn},
+  booktitle={Proceedings of the 3rd International Workshop on Distributed Statistical Computing},
+  volume={124},
+  number={125.10},
+  pages={1--10},
+  year={2003},
+  organization={Vienna, Austria}
+}
+
 @article{Willis_Araújo_Bennett_Figueroa-Rangel_Froyd_Myers_2007,
   title = {How can a knowledge of the past help to conserve the future? Biodiversity Conservation and the relevance of long-term ecological studies},
   author = {Katherine J. Willis and Miguel B. Araújo and Keith D. Bennett and Blanca Figueroa-Rangel and Cynthia A. Froyd and Norman Myers},
@@ -229,4 +251,18 @@
   year = {1987},
   pages = {157–162},
   language = {English}
+}
+
+@article{Syme_2020,
+  author    = {Syme, Don},
+  title     = {The Early History of {F\#}},
+  journal   = {Proceedings of the ACM on Programming Languages},
+  volume    = {4},
+  number    = {HOPL},
+  articleno = {75},
+  pages     = {1--58},
+  month     = {june},
+  year      = {2020},
+  doi       = {10.1145/3386325},
+  url       = {https://doi.org/10.1145/3386325}
 }
