+++ paper/paper.bib
@@ -104,14 +104,6 @@
   language = {en}
 }
 
-@misc{RProvider,
-  title = {RProvider},
-  author = {FsLab},
-  year = {2012},
-  url = {https://github.com/fslaborg/RProvider},
-  note = {Accessed 2026-05-19}
-}
-
 @misc{Purves_2016,
   title = {Filzbach},
   author = {Purves, Drew},
@@ -265,4 +257,16 @@
   year      = {2020},
   doi       = {10.1145/3386325},
   url       = {https://doi.org/10.1145/3386325}
+}
+
+@software{rprovider_contributors_2026_21964392,
+  author       = {RProvider contributors and
+                  Martin, Andrew C.},
+  title        = {RProvider: An F\# Type Provider for R},
+  month        = aug,
+  year         = 2026,
+  publisher    = {Zenodo},
+  version      = {v3.0.0},
+  doi          = {10.5281/zenodo.21964392},
+  url          = {https://doi.org/10.5281/zenodo.21964392},
 }
