### 2026-08-16 02ca6f0
Reorganise stan references
