+++ paper/paper.md
@@ -33,7 +33,7 @@ In contemporary ecology (‘*neo-ecology*’), the 'ecological detective' is an 
 
 *Bristlecone* is a toolkit that enables LTE researchers to utilise the ecological detective approach through: (1) a human-readable, succinct declarative grammar for the ‘ecological detective’ workflow; and (2) implementations of key components required to use the grammar. The toolkit includes a grammar of time that reflects the dating methods used within LTE and neo-ecology; it thus targets neo-ecologists who would like to integrate neo-ecological and LTE time-series for multi-scale analysis.
 
-*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan [@carpenter2017stan], JAGS [@plummer2003jags]) lack dimensional correctness and often require manual construction of temporal concepts. Stan's language, for example, does not support dimensional correctness [@Stan-Development-Team_2025]. *Bristlecone* provides that missing layer within a formalised ecological detective workflow.
+*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan [@carpenter2017stan; @Stan-Development-Team_2025], JAGS [@plummer2003jags]) lack dimensional correctness and often require manual construction of temporal concepts. Stan's language, for example, does not support dimensional correctness. *Bristlecone* provides that missing layer within a formalised ecological detective workflow.
 
 ## State of the field
 
