### 2026-08-16 AndrewIOM/bristlecone#64 @AndrewIOM
Hi @kMutagene , apologies for the delay but I've finally implemented the changes you requested here.

For three of the comments (L14, L58-59, and L135), I have not made any changes, as I believe that the in-text narrative reference format is correct for JOSS, having consulted their journal style guide. They were also auto-generated that way from the bib and md files.

I'll merge the PR I made and close this issue with it, but please re-open if you don't think I addressed any of the points correctly. Many thanks!
