### 2026-08-16 10bbffb
Fixes to paper reviewer comments (#65)
* fix1

* Address comments

* Reorganise stan references
