+++ paper/paper.md
@@ -33,13 +33,13 @@ In contemporary ecology (‘*neo-ecology*’), the 'ecological detective' is an 
 
 *Bristlecone* is a toolkit that enables LTE researchers to utilise the ecological detective approach through: (1) a human-readable, succinct declarative grammar for the ‘ecological detective’ workflow; and (2) implementations of key components required to use the grammar. The toolkit includes a grammar of time that reflects the dating methods used within LTE and neo-ecology; it thus targets neo-ecologists who would like to integrate neo-ecological and LTE time-series for multi-scale analysis.
 
-*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan, JAGS) lack dimensional correctness and often require manual construction of temporal concepts. Stan's language, for example, does not support dimensional correctness [@Stan-Development-Team_2025]. *Bristlecone* provides that missing layer within a formalised ecological detective workflow.
+*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan [@carpenter2017stan; @Stan-Development-Team_2025], JAGS [@plummer2003jags]) lack dimensional correctness and often require manual construction of temporal concepts. Stan's language, for example, does not support dimensional correctness. *Bristlecone* provides that missing layer within a formalised ecological detective workflow.
 
 ## State of the field
 
 The ecological community predominantly utilises R. Within LTE, key statistical approaches include transfer functions, generalised additive models, and canonical correspondence analyses. Proxy-specific mechanistic approaches have also been applied; the landscape models REVEALS [@Sugita_2007] and LOVE [@Sugita_2007_2] simulate taxon-specific biomass from pollen fluxes. However, mechanistic inference from LTE time-series has been limited (e.g., @Jeffers_Bonsall_Watson_Willis_2011). In dendrochronology, (semi-)mechanistic approaches are more prevalent, such as inferring temperature/moisture limits to tree growth using VS-Lite [@TolwinskiWard_2011].
 
-*Bristlecone*'s two unique contributions are: (1) integrating the ecological detective workflow within a single conceptual framework; and (2) providing ecological clarity through dimensional correctness. *Bristlecone* utilises F#’s type system and expressiveness to enforce dimensionally-consistent ecological models and promote human readablity, improving transparency and reproducibility of supplementary material. As the core design relies on F# language features that could not be replicated within R, *Bristlecone* was designed from scratch. However, a compositional approach was sought whereby components may be substituted where beneficial alternatives exist (e.g. optimisers). The F# R type provider [@RProvider] presents an avenue for embedding R libraries and graphics via typed R access.
[…]
