+++ paper/paper.bib
@@ -104,14 +104,6 @@
   language = {en}
 }
 
-@misc{RProvider,
-  title = {RProvider},
-  author = {FsLab},
-  year = {2012},
-  url = {https://github.com/fslaborg/RProvider},
-  note = {Accessed 2026-05-19}
-}
-
 @misc{Purves_2016,
   title = {Filzbach},
   author = {Purves, Drew},
@@ -152,6 +144,28 @@
   month = {Sep},
   journal = {Stan Docs},
   url = {https://mc-stan.org/docs/reference-manual/}
+}
+
+@article{carpenter2017stan,
+  title={{Stan}: A probabilistic programming language},
+  author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+  journal={Journal of Statistical Software},
+  volume={76},
+  number={1},
+  pages={1--32},
+  year={2017},
+  doi={10.18637/jss.v076.i01}
+}
+
+@inproceedings{plummer2003jags,
+  title={{JAGS}: A program for analysis of {B}ayesian graphical models using {G}ibbs sampling},
+  author={Plummer, Martyn},
+  booktitle={Proceedings of the 3rd International Workshop on Distributed Statistical Computing},
+  volume={124},
+  number={125.10},
+  pages={1--10},
+  year={2003},
+  organization={Vienna, Austria}
 }
 
 @article{Willis_Araújo_Bennett_Figueroa-Rangel_Froyd_Myers_2007,
@@ -229,4 +243,30 @@
   year = {1987},
   pages = {157–162},
   language = {English}
+}
+
+@article{Syme_2020,
+  author    = {Syme, Don},
+  title     = {The Early History of {F\#}},
+  journal   = {Proceedings of the ACM on Programming Languages},
+  volume    = {4},
+  number    = {HOPL},
+  articleno = {75},
+  pages     = {1--58},
+  month     = {june},
+  year      = {2020},
+  doi       = {10.1145/3386325},
+  url       = {https://doi.org/10.1145/3386325}
+}
+
+@software{rprovider_contributors_2026_21964392,
+  author       = {RProvider contributors and
+                  Martin, Andrew C.},
+  title        = {RProvider: An F\# Type Provider for R},
+  month        = aug,
+  year         = 2026,
+  publisher    = {Zenodo},
+  version      = {v3.0.0},
+  doi          = {10.5281/zenodo.21964392},
+  url          = {https://doi.org/10.5281/zenodo.21964392},
 }
