### 2026-08-16 #10962 @AndyBunn
Hi @AndrewIOM

V nice paper. See below.

@kdorheim, let me know if I should post this as a linked issue or if it is OK here.

# Review comments: Bristlecone (JOSS submission)

This was a fun read. As a longtime bristlecone pine researcher, I approve of the name and framing! I read *The Ecological Detective* as a student a zillion years ago, and it's wonderful to see it featured so skillfully in an open-source software application. The nod to VS-Lite from Suz Tolwinski-Ward in the State of the field section was a nice touch too, good to see the dendrochronology side of mechanistic modelling represented alongside the LTE framing.

*Note: I'm not an F# user, so my comments below are scoped to the areas the editor asked me to cover (State of the field, Research impact statement, and usability), and I haven't attempted to evaluate the F# code itself for correctness or style.*

## State of the field

1. The claim that Bristlecone fills a gap left by Stan/JAGS on "dimensional correctness" rests on F#'s built-in units-of-measure feature. Could the author clarify what the DSL adds beyond exposing that language feature in ecological vocabulary? This would strengthen the novelty claim.
2. Line 54, "The ecological community predominantly utilises R," seems right to me but has no citation. I recall a paper in an ESA pub supporting this from severl years ago. This is minor, but worth a source or softening.
3. Typo, line 78: "estimation enignes" should read "estimation engines."

## Research impact statement

4. This section isn't a required JOSS element AFAIK (Summary and Statement of need are), so it's a fine addition, but all three cited applications (Jeffers reproduction, Martin 2019/2021, Martin 2024) are the author's own work. Is there any evidence of use by other researchers or groups, issues or discussions from external users, forks with substantive changes, citations? If not, "research impact" might be reframed as "research applications" or similar, since the current title implies broader reach than what's shown.

## Usability

5. Installation, examples, and functionality documentation are all strong. The docs site (https://andrewiom.github.io/bristlecone) has five worked examples with downloadable scripts/notebooks and a full API reference, this exceeds the typical JOSS bar.
6. Tests and CI are present (`tests/Bristlecone.Tests`, GitHub Actions badge, benchmark suite in `benchmarks.md`). Worth confirming the CI badge reflects a currently passing build, not a stale one.
7. Contributing.md is thorough (issue/PR process, code style via fantomas, preference for property tests, docs requirement for new features). No changes needed here.
8. Possible inconsistency to flag for the author: the paper's code example (Software design section, discrete-time Schaefer fish-stock model) uses different API patterns than the README quick-start example (continuous-time, `Model.empty`, `addRateEquation`, `mkContinuous`). Both may be valid, but could the author confirm the paper's code example still runs against the current released version of the software?

This is ready for publication I believe although, as I said, the F# is beyond my ken.
