+++ paper/paper.md
@@ -33,11 +33,11 @@ In contemporary ecology (‘*neo-ecology*’), the 'ecological detective' is an 
 
 *Bristlecone* is a toolkit that enables LTE researchers to utilise the ecological detective approach through: (1) a human-readable, succinct declarative grammar for the ‘ecological detective’ workflow; and (2) implementations of key components required to use the grammar. The toolkit includes a grammar of time that reflects the dating methods used within LTE and neo-ecology; it thus targets neo-ecologists who would like to integrate neo-ecological and LTE time-series for multi-scale analysis.
 
-*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan [@carpenter2017stan; @Stan-Development-Team_2025], JAGS [@plummer2003jags]) lack dimensional correctness and often require manual construction of temporal concepts. Stan's language, for example, does not support dimensional correctness. *Bristlecone* provides that missing layer within a formalised ecological detective workflow.
+*Bristlecone* occupies a distinct niche: although tools exist for data wrangling and inference, *Bristlecone* is a higher-level grammatical layer focused on correctness in model construction from an ecological perspective. General-purpose inference engines (e.g., Stan [@carpenter2017stan; @Stan-Development-Team_2025], JAGS [@plummer2003jags]) do not support dimensional correctness and hold limited built-in temporal concepts. *Bristlecone* embeds units-of-measure within a DSL: ecological states and flows, error structures, and temporal resolutions are expressed as explicit ecological semantics. The DSL enforces not only dimensional consistency but also ecological model structure, for example: rates must be expressed over time; likelihoods correspond to the correct ecological state; and temporal resolutions match between solvers and models. Stocks / flows, parameters, processes, and environmental forcings must thus be expressed in ecological terms. *Bristlecone* thus formalises the ecological detective workflow using ecological semantics.
 
 ## State of the field
 
-The ecological community predominantly utilises R. Within LTE, key statistical approaches include transfer functions, generalised additive models, and canonical correspondence analyses. Proxy-specific mechanistic approaches have also been applied; the landscape models REVEALS [@Sugita_2007] and LOVE [@Sugita_2007_2] simulate taxon-specific biomass from pollen fluxes. However, mechanistic inference from LTE time-series has been limited (e.g., @Jeffers_Bonsall_Watson_Willis_2011). In dendrochronology, (semi-)mechanistic approaches are more prevalent, such as inferring temperature/moisture limits to tree growth using VS-Lite [@TolwinskiWard_2011].
+R is the favoured language of ecologists, increasing in prevalence within high-impact journals from 10.9% in 2013 to 66.9% in 2023 [@Gao_2025]. Within LTE, key statistical approaches include transfer functions, generalised additive models, and canonical correspondence analyses. Proxy-specific mechanistic approaches have also been applied; the landscape models REVEALS [@Sugita_2007] and LOVE [@Sugita_2007_2] simulate taxon-specific biomass from pollen fluxes. However, mechanistic inference from LTE time-series has been limited (e.g., @Jeffers_Bonsall_Watson_Willis_2011). In dendrochronology, (semi-)mechanistic approaches are more prevalent, such as inferring temperature/moisture limits to tree growth using VS-Lite [@TolwinskiWard_2011].
 
[…]
