+++ paper/paper.bib
@@ -270,3 +270,19 @@
   doi          = {10.5281/zenodo.21964392},
   url          = {https://doi.org/10.5281/zenodo.21964392},
 }
+
+@article{Gao_2025,
+    author = {Gao, Meixiang and Ye, Yanyan and Zheng, Ye and Lai, Jiangshan},
+    title = {A comprehensive analysis of R’s application in ecological research from 2008 to 2023},
+    journal = {Journal of Plant Ecology},
+    volume = {18},
+    number = {1},
+    pages = {rtaf010},
+    year = {2025},
+    month = {02},
+    abstract = {The field of ecology has been greatly enhanced by the integration of computational tools and statistical methods, with the programming language R emerging as a pivotal and flexible tool for ecological research. As ecological studies accelerate, understanding the prevalent trends and specific usage patterns of R in recent research is crucial. This study investigated the use of R and its packages in 125 494 scholarly articles published in 40 ecology journals from 2008 to 2023. A total of 52 658 articles (42\%) designated R as their principal analytical tool, demonstrating a steady linear growth in its utilization from 10.3\% in 2008 to 66.9\% in 2023. Twelve R packages, including ‘lme4’, ‘vegan’, ‘nlme’, ‘MuMIn’, ‘ape’, ‘ggplot2’, ‘car’, ‘mgcv’, ‘MASS’, ‘raster’, ‘multcomp’ and ‘lmerTest’, each played a pivotal role in contributing to more than 1000 scholarly articles. The highest usage rate of the 'lme4' package indicates that mixed-effect models have a particularly important role in ecological research, and the application of these models has helped ecologists solve many important scientific problems. Journal-specific package preferences aligned with their scientific domains, while the rise in the average number of R packages per article indicates a trend towards more complex and diverse analytical methods in ecology. Our findings reveal a reciprocal relationship between the development of R and ecological research, underscoring the need for collaboration among quantitative ecologists, R developers and ecologists to further advance both the language and the field. Such collaboration will not only enhance the functionality and versatility of R but also provide robust technical support for the continued progress of ecological research.},
+    issn = {1752-993X},
+    doi = {10.1093/jpe/rtaf010},
+    url = {https://doi.org/10.1093/jpe/rtaf010},
+    eprint = {https://academic.oup.com/jpe/article-pdf/18/1/rtaf010/61458573/rtaf010.pdf},
+}
