### 2026-08-17 7046a79
Changes to address reviewer #2 for JOSS (#67)
* Enforce discrete/continuous eqs at compile-time

* Address reviewer comments in paper

* Fantomas formatting

* Add JOSS button
