### 2026-08-17 #10962 @AndrewIOM
Thank you @AndyBunn for your positive review!

I've worked through the eight points raised and addressed them in the above PR. I've taken the liberty of merging without waiting for a response, but if you think if I didn't fully address any of the points please let me know and I can revise my changes.

Hi @kdorheim, please see the above, and let me know if there is anything further I can do; many thanks!
