### 2026-08-18 #10962 @kdorheim
Thank you @AndrewIOM for your quick response to the reviewer feedback!

@kMutagene and @AndyBunn, if you could please take a look at the changes and assess whether, with these modifications, you feel comfortable completing the reviewer checklist, we can wrap up the review process and move on to the next steps toward publication.

If you feel that further revisions or iterations are necessary, we should, of course, continue the review process, as that is an important part of the JOSS process. I’m happy to facilitate any further review or discussion if needed.
