### 2026-08-19 AndrewIOM/bristlecone#64 @kMutagene
Still got one weird looking one here where only the number is a link but maybe thats how the citation style handles boockchapters, anyways not trying to be pedantic. Thank your for tacking care of the rest
