### 2026-08-19 #10962 @kMutagene
Hi @kdorheim the recent revisions have further improved what was already a strong manuscript, and I wholeheartedly recommend this work for publication.

The only checklist item for which I had some uncertainty was `Development history and open-source practice > Collaborative effort`. My understanding is that this submission represents the maturation of software originally developed and used in the author's scientific work into a well-documented, and citable software library for future research. I do not see the fact that this development is driven primarily by a single researcher as a concern in this case, especially given the high quality of both the codebase and its documentation. I have therefore marked this item as satisfied.
