### 2026-08-24 #10962 @kdorheim
Hi @kMutagene, thank you for participating in the JOSS review process. Yes, I wholeheartedly agree with your assessment of the collaborative effort. @AndyBunn, do you feel that the author's revisions address your feedback? Important note our other reviewer, @kMutagene, is a F# expert, and their review focused on that.
