### 2026-08-24 #10962 @AndyBunn
Thanks for cajoling me. Hope it wasn't too painful.
