### 2026-08-27 215d911
Update paper ref formatting
