### 2026-08-27 1c91f28
Final paper adjustments for JOSS (#69)
* Update paper ref formatting

* Add citation file and release 3.1.2 for joss article archive
