+++ paper/paper.bib
@@ -1,288 +1,263 @@
 @article{Willis_Bailey_Bhagwat_Birks_2010,
-  title = {Biodiversity baselines, thresholds and resilience: testing predictions and assumptions using palaeoecological data},
-  volume = {25},
-  doi = {10.1016/j.tree.2010.07.006},
-  abstractnote = {Fossil records are replete with examples of long-term biotic responses to past climate change. One particularly useful set of records are those preserved in lake and marine sediments, recording both climate changes and corresponding biotic responses. Recently there has been increasing focus on the need for conservation of ecological and evolutionary processes in the face of climate change. We review key areas where palaeoecological archives contribute to this conservation goal, namely: (i) determination of rates and nature of biodiversity response to climate change; (ii) climate processes responsible for ecological thresholds; (iii) identification of ecological resilience to climate change; and (iv) management of novel ecosystems. We stress the importance of long-term palaeoecological data in fully understanding contemporary and future biotic responses.},
-  number = {10},
-  journal = {Trends in Ecology & Evolution},
-  author = {K.J. Willis and R.M. Bailey and S.A. Bhagwat and H.J.B. Birks},
-  year = {2010},
-  pages = {583–591},
-  language = {en}
+	title        = {Biodiversity baselines, thresholds and resilience: testing predictions and assumptions using palaeoecological data},
+	author       = {K. J. Willis and R .M. Bailey and S .A. Bhagwat and H. J. B. Birks},
+	year         = 2010,
+	journal      = {Trends in Ecology & Evolution},
+	volume       = 25,
+	number       = 10,
+	pages        = {583–591},
+	doi          = {10.1016/j.tree.2010.07.006},
+	abstractnote = {Fossil records are replete with examples of long-term biotic responses to past climate change. One particularly useful set of records are those preserved in lake and marine sediments, recording both climate changes and corresponding biotic responses. Recently there has been increasing focus on the need for conservation of ecological and evolutionary processes in the face of climate change. We review key areas where palaeoecological archives contribute to this conservation goal, namely: (i) determination of rates and nature of biodiversity response to climate change; (ii) climate processes responsible for ecological thresholds; (iii) identification of ecological resilience to climate change; and (iv) management of novel ecosystems. We stress the importance of long-term palaeoecological data in fully understanding contemporary and future biotic responses.},
+	language     = {en}
 }
-
 @book{Hilborn_Mangel_1997,
-  title = {The Ecological Detective: Confronting models with data},
-  author = {Hilborn, Ray and Mangel, Marc},
-  year = {1997},
-  publisher = {Princeton University Press},
-  place = {Princeton, New Jersey}
+	title        = {The Ecological Detective: Confronting models with data},
+	author       = {Hilborn, Ray and Mangel, Marc},
+	year         = 1997,
+	publisher    = {Princeton University Press},
+	place        = {Princeton, New Jersey}
 }
-
 @article{McLauchlan_Gerhart_2017,
-  title = {Centennial-scale reductions in nitrogen availability in temperate forests of the United States.},
-  volume = {7},
-  doi = {10.1038/s41598-017-08170-z},
-  number = {1},
-  journal = {Scientific Reports},
-  publisher = {Springer Science and Business Media LLC},
-  author = {K. K. McLauchlan and L. M. Gerhart and J. J. Battles and J. M. Craine and A. J. Elmore and P. E. Higuera and M. C. Mack and B. E. McNeil and D. M. Nelson and N. Pederson and S. S. Perakis},
-  year = {2017},
-  pages = {7856},
-  language = {English}
[…]
