### 2026-08-27 #10962 @AndrewIOM
Hi @kdorheim, thanks. I've made the fixes you've requested and completed the checklist. Here is the Zenodo archive reference for the newly released version:

Martin, A. C. (2026). Bristlecone: an F# library for the long-term ecological detective (Version v3.1.2) [Computer software]. Zenodo. https://doi.org/10.5281/zenodo.22132894

On the line-by-line changes, I fixed all of the 5x reference-related issues, but could not fix the spacing issues. The spaces are all single spaces in the markdown, but are being rendered larger by the pandoc formatting stretching to fill the width, which I don't seem to have control over. I hope that's OK!

Many thanks
