### 2026-08-27 #10962 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
