### 2026-09-03 #10962 @kthyng
@AndrewIOM Paper comments:
- Page 1: “the long view is integral to comprehensively understand ecosystem responses” should be “the long view is integral to comprehensively understanding ecosystem responses.”
- Page 2: “promote human readablity” contains a typo: “readablity” should be “readability.”
- Page 5: “I thank support from Dr Elizabeth Jeffers and Prof Marc Macias-Fauria” should be “I thank Dr Elizabeth Jeffers and Prof Marc Macias-Fauria for their support.” or similar rewriting.
- Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example:
  - Martin (2019) reference: “western siberian tundra” should be “western Siberian tundra.”
  - Martin et al. (2024) reference: “The arctic holocene biodiversity database” should capitalize the proper names: “The Arctic Holocene biodiversity database.”
  - Martin et al. (2021) reference: “arctic shrubs” should be “Arctic shrubs.”
  - McLauchlan et al. (2017) reference: “the united states” should be “the United States.”
