### 2026-09-04 fcfd81b
Final paper typo / ref casing changes
