+++ paper/paper.bib
@@ -11,14 +11,14 @@
 	language     = {en}
 }
 @book{Hilborn_Mangel_1997,
-	title        = {The Ecological Detective: Confronting models with data},
+	title        = {The ecological detective: confronting models with data},
 	author       = {Hilborn, Ray and Mangel, Marc},
 	year         = 1997,
 	publisher    = {Princeton University Press},
 	place        = {Princeton, New Jersey}
 }
 @article{McLauchlan_Gerhart_2017,
-	title        = {Centennial-scale reductions in nitrogen availability in temperate forests of the United States.},
+	title        = {Centennial-scale reductions in nitrogen availability in temperate forests of the {United States}.},
 	author       = {K. K. McLauchlan and L. M. Gerhart and J. J. Battles and J. M. Craine and A. J. Elmore and P. E. Higuera and M. C. Mack and B. E. McNeil and D. M. Nelson and N. Pederson and S. S. Perakis},
 	year         = 2017,
 	journal      = {Scientific Reports},
@@ -56,7 +56,7 @@
 	language     = {en}
 }
 @article{Sugita_2007,
-	title        = {Theory of quantitative reconstruction of vegetation {I}: pollen from large sites REVEALS regional vegetation composition:},
+	title        = {Theory of quantitative reconstruction of vegetation {I}: pollen from large sites {REVEALS} regional vegetation composition:},
 	author       = {Shinya Sugita},
 	year         = 2007,
 	journal      = {The Holocene},
@@ -70,7 +70,7 @@
 	language     = {English}
 }
 @article{Sugita_2007_2,
-	title        = {Theory of quantitative reconstruction of vegetation II: all you need is LOVE:},
+	title        = {Theory of quantitative reconstruction of vegetation {II}: all you need is {LOVE}},
 	author       = {Shinya Sugita},
 	year         = 2007,
 	journal      = {The Holocene},
@@ -105,7 +105,7 @@
 	url          = {https://github.com/predictionmachines/Filzbach/tree/master/filzbach}
 }
 @misc{baydin2015diffsharpautomaticdifferentiationlibrary,
-	title        = {DiffSharp: Automatic Differentiation Library},
+	title        = {{DiffSharp}: Automatic Differentiation Library},
 	author       = {Atilim Gunes Baydin and Barak A. Pearlmutter and Jeffrey Mark Siskind},
 	year         = 2015,
 	url          = {https://arxiv.org/abs/1511.07727},
@@ -114,7 +114,7 @@
 	primaryclass = {cs.MS}
 }
 @article{Jeffers_Bonsall_Watson_Willis_2011,
-	title        = {Climate change impacts on ecosystem functioning: evidence from an Empetrum heathland},
+	title        = {Climate change impacts on ecosystem functioning: evidence from an {Empetrum} heathland},
 	author       = {Elizabeth S. Jeffers and Michael B. Bonsall and Jenny E. Watson and Katherine J. Willis},
 	year         = 2011,
 	month        = {Oct},
@@ -153,7 +153,7 @@
 	organization = {Vienna, Austria}
 }
 @article{Willis_Araújo_Bennett_Figueroa-Rangel_Froyd_Myers_2007,
-	title        = {How can a knowledge of the past help to conserve the future? Biodiversity Conservation and the relevance of long-term ecological studies},
+	title        = {How can a knowledge of the past help to conserve the future? Biodiversity conservation and the relevance of long-term ecological studies},
 	author       = {K. J. Willis and Miguel B. Araújo and Keith D. Bennett and Blanca Figueroa-Rangel and Cynthia A. Froyd and Norman Myers},
 	year         = 2007,
[…]
