### 2026-09-04 d98bc20
Final paper typo / ref casing changes (#70)
